+++ src/sb_counter.c
@@ -52,8 +52,8 @@ void sb_counters_done(void)
 
 static void sb_counters_merge(sb_counters_t dst)
 {
-  for (size_t t = 0; t < SB_CNT_MAX; t++)
-    for (size_t i = 0; i < sb_globals.threads; i++)
+  for (unsigned int t = 0; t < SB_CNT_MAX; t++)
+    for (unsigned int i = 0; i < sb_globals.threads; i++)
       dst[t] += sb_counter_val(i, t);
 }
