+++ src/tests/fileio/sb_fileio.c
@@ -356,8 +356,8 @@ int file_prepare(void)
       log_text(LOG_FATAL,
                "Size of file '%s' is %sB, but at least %sB is expected.",
                file_name,
-               sb_print_value_size(ss1, sizeof(ss1), buf.st_size),
-               sb_print_value_size(ss2, sizeof(ss2), file_size));
+               sb_print_value_size(ss1, sizeof(ss1), (double)buf.st_size),
+               sb_print_value_size(ss2, sizeof(ss2), (double)file_size));
       log_text(LOG_WARNING,
                "Did you run 'prepare' with different --file-total-size or "
                "--file-num values?");
@@ -590,7 +590,7 @@ int file_execute_event(sb_event_t *sb_req, int thread_id)
     log_text(LOG_FATAL, "Incorrect file id in request: %u", file_req->file_id);
     return 1;
   }
-  if (file_req->pos + file_req->size > file_size)
+  if (file_req->pos + (intptr_t)file_req->size > file_size)
   {
     log_text(LOG_FATAL, "I/O request exceeds file size. "
              "file id: %d file size: %lld req offset: %lld req size: %lld",
@@ -608,7 +608,7 @@ int file_execute_event(sb_event_t *sb_req, int thread_id)
 
       /* Store checksum and offset in a buffer when in validation mode */
       if (sb_globals.validate)
-        file_fill_buffer(per_thread[thread_id].buffer, file_req->size, file_req->pos);
+        file_fill_buffer(per_thread[thread_id].buffer, (unsigned int)file_req->size, file_req->pos);
                          
       if(file_pwrite(file_req->file_id, per_thread[thread_id].buffer,
                      file_req->size, file_req->pos, thread_id)
@@ -643,7 +643,7 @@ int file_execute_event(sb_event_t *sb_req, int thread_id)
 
       /* Validate block if run with validation enabled */
       if (sb_globals.validate &&
-          file_validate_buffer(per_thread[thread_id].buffer, file_req->size, file_req->pos))
+          file_validate_buffer(per_thread[thread_id].buffer, (unsigned int)file_req->size, file_req->pos))
       {
         log_text(LOG_FATAL,
           "Validation failed on file " FD_FMT ", block offset %lld, exiting...",
@@ -692,16 +692,16 @@ void file_print_mode(void)
 
   print_file_extra_flags();
   log_text(LOG_NOTICE, "%d files, %sB each", num_files,
-           sb_print_value_size(sizestr, sizeof(sizestr), file_size));
+           sb_print_value_size(sizestr, sizeof(sizestr), (double)file_size));
   log_text(LOG_NOTICE, "%sB total file size",
            sb_print_value_size(sizestr, sizeof(sizestr),
-                               file_size * num_files));
+                               (double)(file_size * num_files)));
   log_text(LOG_NOTICE, "Block size %sB",
-           sb_print_value_size(sizestr, sizeof(sizestr), file_block_size));
+           sb_print_value_size(sizestr, sizeof(sizestr), (double)file_block_size));
   if (file_merged_requests > 0)
     log_text(LOG_NOTICE, "Merging requests up to %sB for sequential IO.",
              sb_print_value_size(sizestr, sizeof(sizestr),
-                                 file_request_size));
+                                 (double)file_request_size));
 
   switch (test_mode)
   {
@@ -1606,7 +1606,7 @@ int parse_arguments(void)
     return 1;
   }
   
-  file_block_size = sb_get_value_size("file-block-size");
+  file_block_size = (int)sb_get_value_size("file-block-size");
   if (file_block_size <= 0)
   {
     log_text(LOG_FATAL, "Invalid value for file-block-size: %d.",
